+++ services/web-app/lib/github.js
@@ -291,7 +291,8 @@ const mergeInheritedAssets = (assets = [], inheritAssets = []) => {
       inheritableProperties.forEach((property) => {
         const inheritProperty = get(inheritAsset, `content.${property}`)
 
-        if (!get(asset, `content.${property}`) && inheritProperty) {
+        // check for undefined properties, as well as empty arrays for tags
+        if (isEmpty(get(asset, `content.${property}`)) && inheritProperty) {
           set(asset, `content.${property}`, inheritProperty)
         }
       })
@@ -299,6 +300,20 @@ const mergeInheritedAssets = (assets = [], inheritAssets = []) => {
 
     return asset
   })
+}
+
+/**
+ * Ensures an asset has default properties if not set
+ * @param {import('../typedefs').AssetContent} assetContent
+ * @returns {import('../typedefs').AssetContent}
+ */
+const mergeAssetContentDefaults = (assetContent = {}) => {
+  return {
+    ...assetContent,
+    noIndex: !!assetContent.noIndex && process.env.INDEX_ALL !== '1', // default to false if not specified
+    framework: assetContent?.framework ?? 'design-only',
+    tags: assetContent?.tags ?? []
+  }
 }
 
 /**
@@ -854,11 +869,10 @@ const getLibraryAssets = async (params = {}) => {
         return {
           params: libraryParams,
           response,
-          content: {
+          content: mergeAssetContentDefaults({
             id: assetKey,
-            ...asset,
-            noIndex: !!asset.noIndex && process.env.INDEX_ALL !== '1' // default to false if not specified
-          }
+            ...asset
+          })
         }
       })
     )