+++ services/web-app/pages/libraries/[host]/[org]/[repo]/[library]/[ref]/assets/[asset]/index.js
@@ -148,11 +148,14 @@ const Asset = ({ libraryData, overviewMdxSource, params }) => {
     }
   ]
 
-  if (libraryData.content.demoLinks) {
-    pageNavItems.push({
-      title: 'Demo links',
-      id: 'demo-links'
-    })
+  if (assetData.content.demoLinks) {
+    setPageNavItems([
+      ...pageNavItems,
+      {
+        title: 'Demo links',
+        id: 'demo-links'
+      }
+    ])
   }
 
   const githubRepoUrl = `https://${assetData.params.host}/${assetData.params.org}/${assetData.params.repo}`
@@ -179,7 +182,7 @@ const Asset = ({ libraryData, overviewMdxSource, params }) => {
           href={`/libraries/${frameworks.params.library}/${params.ref}/assets/${params.asset}`}
           passHref
         >
-          <CarbonLink size="lg">{frameworks.framework}</CarbonLink>
+          <CarbonLink size="lg">{framework[frameworks.framework]?.name}</CarbonLink>
         </Link>
       </>
     ))
@@ -276,7 +279,7 @@ const Asset = ({ libraryData, overviewMdxSource, params }) => {
                       lg={4}
                     >
                       <dt className={dashboardStyles.label}>Other frameworks</dt>
-                      <dd className={clsx(dashboardStyles.meta, styles['other-frameworks'])}>
+                      <dd className={dashboardStyles.meta}>
                         {otherFrameworks.length > 0 ? otherFrameworkLinks : '–'}
                       </dd>
                     </Column>
@@ -344,11 +347,11 @@ const Asset = ({ libraryData, overviewMdxSource, params }) => {
                   )}
                 </DashboardItem>
               </Column>
-              {libraryData.content.demoLinks && (
+              {assetData.content.demoLinks && (
                 <Column sm={4} md={8} lg={8}>
                   <section id="demo-links">
                     <H2>Demo links</H2>
-                    <DemoLinks links={[...get(libraryData, 'content.demoLinks', [])]} />
+                    <DemoLinks links={[...get(assetData, 'content.demoLinks', [])]} />
                   </section>
                 </Column>
               )}